Update chosen - move common items into abstract class.
